+++ src/web/FloraWeb/Pages/Templates/Pages/Categories/Show.hs
@@ -14,4 +14,4 @@ showCategory :: Category -> Vector PackageInfo -> FloraHTML
 showCategory Category{name, synopsis} packagesInfo = do
   div_ [class_ "container"] $ do
     presentationHeader name synopsis (fromIntegral $ V.length packagesInfo)
-    packageListing packagesInfo
+    packageListing Nothing packagesInfo