+++ src/web/FloraWeb/Pages/Templates/Pages/Search.hs
@@ -1,7 +1,6 @@
 module FloraWeb.Pages.Templates.Pages.Search where
 
 import Control.Monad (when)
-import Data.Foldable (forM_)
 import Data.Positive
 import Data.Text (Text)
 import Data.Text.Display (display)
@@ -11,7 +10,6 @@ import Lucid
 import Flora.Model.Package (Namespace, PackageInfo (..))
 import Flora.Search (SearchAction (..))
 import FloraWeb.Components.PackageListHeader (presentationHeader)
-import FloraWeb.Components.PackageListItem
 import FloraWeb.Components.PaginationNav (paginationNav)
 import FloraWeb.Pages.Templates
 import FloraWeb.Pages.Templates.Packages (packageListing)
@@ -20,14 +18,14 @@ showAllPackages :: Word -> Positive Word -> Vector PackageInfo -> FloraHTML
 showAllPackages count currentPage packagesInfo = do
   div_ [class_ "container"] $ do
     presentationHeader "Packages" "" count
-    div_ [class_ ""] $ packageListing packagesInfo
+    div_ [class_ ""] $ packageListing Nothing packagesInfo
     paginationNav count currentPage ListAllPackages
 
 showAllPackagesInNamespace :: Namespace -> Word -> Positive Word -> Vector PackageInfo -> FloraHTML
 showAllPackagesInNamespace namespace count currentPage packagesInfo = do
   div_ [class_ "container"] $ do
     presentationHeader (display namespace) "" count
-    div_ [class_ ""] $ packageListing packagesInfo
+    div_ [class_ ""] $ packageListing Nothing packagesInfo
     paginationNav count currentPage (ListAllPackagesInNamespace namespace)
 
 showResults
@@ -42,9 +40,6 @@ showResults
 showResults searchString count currentPage exactMatches results = do
   div_ [class_ "container"] $ do
     presentationHeader searchString "" count
-    forM_ exactMatches $ \em -> do
-      div_ [class_ "exact-match"] $
-        packageListItem (em.namespace, em.name, em.synopsis, em.version, em.license)
-    div_ [class_ ""] $ packageListing results
+    packageListing (Just exactMatches) results
     when (count > 30) $
       paginationNav count currentPage (SearchPackages searchString)