+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add a page on namespaces in the documentation ([#451](https://github.com/flora-pm/flora-server/pull/451))
 * Add initial support for hosting package tarballs ([#452](https://github.com/flora-pm/flora-server/pull/452))
 * Show depended on components in dependencies page ([#464](https://github.com/flora-pm/flora-server/pull/464))
+* Add search bar for reverse dependencies ([#476](https://github.com/flora-pm/flora-server/pull/476))
 
 ## 1.0.13 -- 2023-09-17
 * Exclude deprecated releases from latest versions and search ([#373](https://github.com/flora-pm/flora-server/pull/373))